print load test data in respect to facts in DB

+++ specs.wdio/testapp/server.ts
@@ -11,8 +11,10 @@ createServer({ transportDriver: http }).then((server) => {
 // TODO: this is only needed for the karma test, can be removed once all tests are migrated to wdio
 http.createServer(async (req, res) => {
   if (req.url === '/deleteFacts') {
-    await pgPool.query('TRUNCATE facts;');
-    console.log('TRUNCATE facts done');
+    if (process.env['NO_DB_TRUNCATE_ON_TEST'] !== 'true') {
+      await pgPool.query('TRUNCATE facts;');
+      console.log('TRUNCATE facts done');
+    }
   }
 
   res.writeHead(200);
